Make OptionsManager const, fix config_path, comment new methods and minor fixes

+++ cvmfs/publish/repository.cc
@@ -808,9 +808,10 @@ void Publisher::ExitShell() {
   std::string session_pid_tmp = session_dir + "/session_pid";
   std::string session_pid;
   int fd_session_pid = open(session_pid_tmp.c_str(), O_RDONLY);
+  if (fd_session_pid < 0) throw EPublish("session pid cannot be retrieved");
   SafeReadToString(fd_session_pid, &session_pid);
 
-  pid_t pid_child = stoi(session_pid);
+  pid_t pid_child = String2Uint64(session_pid);
   kill(pid_child, SIGKILL);
 }
